Use brackets.min.css instead of brackets.less
At the moment *.less files are copied over but also compiled to *.css files at build time.
So we can reference the *.css one directly in index.html
This has also the side effect to speed up a bit the testsuite, since there isn't the needs to recompile them for every window opened.